- Moved eventTime classes to their own package
