+++ src/Supermarket/eventTime/AddPersonTime.java
@@ -1,10 +1,11 @@
-package Supermarket;
+package Supermarket.eventTime;
 
+import Supermarket.SupermarketState;
 import random.UniformRandomStream;
 
 import java.util.Date;
 
-class AddPersonTime extends EventTime {
+public class AddPersonTime extends EventTime {
 
     public AddPersonTime(SupermarketState supermarketState) {
         super(supermarketState, new UniformRandomStream(100, 5000, new Date().getTime()));
