+++ src/Supermarket/eventTime/ByShitTime.java
@@ -1,10 +1,11 @@
-package Supermarket;
+package Supermarket.eventTime;
 
+import Supermarket.SupermarketState;
 import random.UniformRandomStream;
 
 import java.util.Date;
 
-class ByShitTime extends EventTime {
+public class ByShitTime extends EventTime {
 
     public ByShitTime(SupermarketState supermarketState) {
         super(supermarketState, new UniformRandomStream(1000, 20000, new Date().getTime()));
