+++ src/Supermarket/eventTime/ServeTime.java
@@ -1,10 +1,12 @@
-package Supermarket;
+package Supermarket.eventTime;
 
+import Supermarket.SupermarketState;
+import Supermarket.eventTime.EventTime;
 import random.UniformRandomStream;
 
 import java.util.Date;
 
-class ServeTime extends EventTime {
+public class ServeTime extends EventTime {
 
     public ServeTime(SupermarketState supermarketState) {
         super(supermarketState, new UniformRandomStream(500, 10000, new Date().getTime()));
